pyproject.toml
[tool.poetry]
name = "apibara"
version = "0.1.1"
description = "Apibara cliend SDK. Build web3-powered applications."
authors = ["Francesco Ceccon <francesco@apibara.com>"]

[tool.poetry.scripts]
apibara = "apibara.cli:cli"

[tool.poetry.dependencies]
python = "^3.9"
grpcio = "^1.47.0"
grpcio-tools = "^1.47.0"
eth-hash = {extras = ["pysha3"], version = "^0.3.2"}
aiochannel = "^1.1.1"
click = "^8.1.3"
click-help-colors = "^0.9.1"
aiohttp = "^3.8.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.3.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
